% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GettingStarted.R
\name{doSomeADLSOperations}
\alias{doSomeADLSOperations}
\title{Main function that displays some basic ADLS operations.}
\usage{
doSomeADLSOperations()
}
\description{
Main function that displays some basic ADLS operations.
}
