% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GettingStarted.R
\name{getAzureActiveContext}
\alias{getAzureActiveContext}
\title{Create a new `azureActiveContext` using the config.json.}
\usage{
getAzureActiveContext()
}
\value{
`azureActiveContext` object
}
\description{
SAMPLE config_clientcredentials.json:
}
\details{
{
"authType": "ClientCredential",
"resource": "https://datalake.azure.net/",
"tenantID": "72f988bf-blah-41af-blah-2d7cd011blah",
"clientID": "1d604733-blah-4b37-blah-98fca981blah",
"authKey": "zTw5blah+IN+yIblahrKv2K8dM2/BLah4FogBLAH/ME=",
"azureDataLakeAccount": "azuresmrtestadls"
}

SAMPLE config_devicecode.json:

{
"authType": "DeviceCode",
"resource": "https://datalake.azure.net/",
"tenantID": "72fblahf-blah-41af-blah-2d7cBLAHdb47",
"clientID": "2dblah05-blah-4e0a-blah-ae4d2BLAH5df",
}
}
